Move tap animation logic in to the menu-open-button
